dispatcher/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
PyJWT==2.8.0
httpx==0.26.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6
prometheus-client==0.19.0
pytest==7.4.4
pytest-asyncio==0.23.3
